+++ implementation_plan.md
@@ -1,8 +1,8 @@
 # Team Responsibilities for Classes
 
-- **Taha** : `entities.ItemData`, `ItemManager` *(end of Oct 29)*
+- **Taha** : `entities.items.ItemData`, `ItemManager` *(end of Oct 29)*
 - **Chris** : `Event` & `CombatEvent` & `EnemyData` & `QuestionsData`, `CombatInteractor`, `EventManager` *(end of Oct 29)*
-- **Natasha** : `entities.PlayerData`, `entities.PlayerInteractor` *(end of Oct 29)*
+- **Natasha** : `entities.player.PlayerData`, `entities.player.PlayerInteractor` *(end of Oct 29)*
 - **Alaina** : `LoadInteractor`, `<LoadInterface>`, `SaveInteractor`, `<SaveInterface>`
 - **Josh** : `usecases.StorylineInteractor`, `<StorylineInterface>`
 - **Ludan** : `ViewController`, `View`
