+++ src/main/java/UI/UIFacade.java
@@ -3,7 +3,10 @@ package UI;
 import controller.CombatController;
 import controller.LoginController;
 import controller.StorylineController;
-import entities.*;
+import entities.events.EventManager;
+import entities.items.ItemData;
+import entities.player.PlayerData;
+import entities.player.PlayerInteractor;
 import presenter.CombatPresenter;
 import presenter.LoginPresenter;
 import presenter.StorylinePresenter;
