+++ src/main/java/entities/events/CombatEvent.java
@@ -1,4 +1,4 @@
-package entities;
+package entities.events;
 
 import java.util.List;
 import java.util.Random;
