+++ src/main/java/entities/events/Event.java
@@ -1,4 +1,4 @@
-package entities;
+package entities.events;
 
 import java.util.List;
