+++ src/main/java/entities/player/PlayerInteractor.java
@@ -1,16 +1,18 @@
-package entities;  // relevant imports
+package entities.player;  // relevant imports
+
+import entities.items.ItemData;
 
 import java.util.ArrayList;
 
 /**
  * PlayerInteractor class
- * Responsible for updating the entities.PlayerData class.
+ * Responsible for updating the entities.player.PlayerData class.
  */
 @SuppressWarnings("all")
 public class PlayerInteractor {
 
     /**
-     * The entities.PlayerData object that this class will interact with.
+     * The entities.player.PlayerData object that this class will interact with.
      */
     PlayerData player;
 
@@ -23,8 +25,8 @@ public class PlayerInteractor {
     }
 
     /**
-     * Updates the entities.PlayerData object that this class interacts with.
-     * @param p the new entities.PlayerData object.
+     * Updates the entities.player.PlayerData object that this class interacts with.
+     * @param p the new entities.player.PlayerData object.
      */
     public void updatePlayer(PlayerData p) {
         player = p;
@@ -32,7 +34,7 @@ public class PlayerInteractor {
 
     /**
      * Updates the player's attack power.
-     * @param attackPower the value to be added to the entities.PlayerData object.
+     * @param attackPower the value to be added to the entities.player.PlayerData object.
      */
     public void updateAttackPower(int attackPower) {
         player.setAttackPower(player.getAttackPower() + attackPower);
@@ -40,7 +42,7 @@ public class PlayerInteractor {
 
     /**
      * Returns the player's attack power.
-     * @return the entities.PlayerData object's attack power.
+     * @return the entities.player.PlayerData object's attack power.
      */
     public int getAttackPower() {
         return player.getAttackPower();
@@ -54,34 +56,34 @@ public class PlayerInteractor {
     }
 
     /**
-     * Returns what event id is stored in the entities.PlayerData object.
-     * @return The current event stored in the entities.PlayerData object.
+     * Returns what event id is stored in the entities.player.PlayerData object.
+     * @return The current event stored in the entities.player.PlayerData object.
      */
     public int getPlayerEventID() {
         return player.getEventID();
     }
 
     /**
-     * Returns the username of the entities.PlayerData object.
-     * @return The username of the entities.PlayerData object.
+     * Returns the username of the entities.player.PlayerData object.
+     * @return The username of the entities.player.PlayerData object.
      */
     public String getPlayerUsername() {
         return player.getUsername();
     }
 
     /**
-     * Adds an entities.ItemData object to the entities.PlayerData object's inventory.
-     * Creates a new list of entities.ItemData if there is not already one.
-     * @param item The entities.ItemData object that is added to the entities.PlayerData's inventory.
+     * Adds an entities.items.ItemData object to the entities.player.PlayerData object's inventory.
+     * Creates a new list of entities.items.ItemData if there is not already one.
+     * @param item The entities.items.ItemData object that is added to the entities.player.PlayerData's inventory.
      */
     public void addToInventory(ItemData item) {
         player.addToInventory(item);
     }
 
     /**
-     * Removes an entities.ItemData object from the entities.PlayerData object's inventory.
-     * @param item The entities.ItemData object that is removed from the inventory.
-     * @return removed Whether the entities.ItemData object was removed.
+     * Removes an entities.items.ItemData object from the entities.player.PlayerData object's inventory.
+     * @param item The entities.items.ItemData object that is removed from the inventory.
+     * @return removed Whether the entities.items.ItemData object was removed.
      */
     public boolean removeFromInventory(ItemData item) {
         return player.removeFromInventory(item);
@@ -89,7 +91,7 @@ public class PlayerInteractor {
 
     /**
      * Returns the number of items of a type in a list.
-     * @param item The entities.ItemData object that we count in the inventory.
+     * @param item The entities.items.ItemData object that we count in the inventory.
      * @return the number of instances of the item type in the inventory.
      */
     public int itemCount(ItemData item) {
