+++ src/main/java/usecases/Loader.java
@@ -1,6 +1,6 @@
 package usecases;
-import entities.PlayerData;
 
+import entities.player.PlayerData;
 
 import java.io.IOException;
